epedidos-master/requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
PyJWT==1.4.0
wheel==0.24.0
